P2c2: extract afxdp/tx/drain.rs (drain dispatch + queue-bound + pending helpers, ~600 LOC)
Extracts the drain dispatch cluster from tx/mod.rs into sibling tx/drain.rs:
- 5 pub fns: pending_tx_capacity, bound_pending_tx_local,
  bound_pending_tx_prepared, drain_pending_tx,
  drain_pending_tx_local_owner. All bumped from pub(super) to
  pub(in crate::afxdp); facade pub(super) re-export from tx/mod.rs
  preserves sibling reach (worker.rs, frame_tx.rs, afxdp.rs).
- 6 file-private drain helpers: drop_cos_bound_prepared_leftovers,
  partition_cos_bound_local_with_rescue, drop_cos_bound_local_leftovers,
  binding_has_pending_tx_work, ingest_cos_pending_tx,
  ingest_cos_pending_tx_with_provenance.
- 3 pending-queue helpers (file-private): process_pending_queue_in_place,
  take_pending_tx_requests, restore_pending_tx_requests.
- 4 constants: COS_GUARANTEE_VISIT_NS, COS_GUARANTEE_QUANTUM_MIN/MAX_BYTES,
  COS_SURPLUS_ROUND_QUANTUM_BYTES — re-exported pub(in crate::afxdp)
  from tx/mod.rs since cos/queue_service.rs imports them via
  crate::afxdp::tx::*.

Tests for partition_cos_bound_local_with_rescue (2) and
process_pending_queue_in_place (1) moved to drain.rs::tests since
their target fns are file-private in drain.rs.

cos_classify/enqueue cluster + cached-selection (CoSTxSelection,
resolve_cached_cos_tx_selection) and cos_queue_dscp_rewrite stay in
tx/mod.rs for P2d.

drain.rs uses 'use super::*;' for now to inherit tx/mod.rs's wide
import surface (the moved bodies need many sibling/parent helpers).
The explicit import block is deferred to a future cleanup PR per the
plan's stated risk — round-1 reviewer can verify it doesn't pull in
anything dead.

cargo build --bins clean (88 warnings, all pre-existing dead-code; 0
unused-imports new). cargo test --bins 865/0/2 (rolling baseline).

Plan: docs/pr/p2c2-tx-drain/plan.md (v3.1, addressing Codex r1+r2+r3 findings).
